MultiLLM Analysis
Added some changes to the ResultsSummary and dropped the calculations of correlation with TAOM (It's not needed). Need to write method that can loop over each of the directories and perform the calculations, and gather the statistics from each directory for a comparison plot. You want statistics versus Pratio. Also need to investigate why the fitted dnu distribution is bimodal when the measured spectral data is clearly not bimodal and the dnu taken from the data is also clearly not bimodal. Suspect the fitting span is too wide.